Add required changes. Styles now use theme colors and are in a different file.
